Adjust block data source
There has been a change to the serialization of `Leaf` that removes the ability to
deserialize or serialize block data within the `Leaf`. As a result, block size information
is missing.

To fix this issue, we need the `BlockQueryData` in order to determine details about the
`Payload` that are needed for this service. Additionally, the `Leaf` is still needed. So
we need to be able to retrieve two separate `Stream`s, one for the `Leaf`s, and one for
the `BlockQueryData`, and then to combine those entries.  As a result, a solution is needed
to be able to retrieve these entries, and ensure that we are matching them at the correct
offsets.