Add a retry with backoff for initial block height retrieval
When the node validator service is first launching it attempts to retrieve the current
block height from the query node.  If it fails the program cannot continue and it panics
in order to die violently.  However, it need not give up after a single attempt and can
instead make a few attempts with a backoff in order to ensure that it really is a failure
case and not just a temporary failure.

This commit adds a retry loop for 4 attempts with a decaying delay between requests.