Replace sequencer version in espresso-dev-node
The espresso-dev-node that is utilized to simply spinning up a simple testable Espresso
network stack is set to use the `MarketplaceVersion` by default.  This version is not
released and is not planned to be released at this time.  Instead, we'd prefer to use the
currently released or upcoming versions of the Sequencer in order to match real world
testing conditions.

This commit replaces the `MarketplaceVersion` of the sequencer in
`espresso-dev-node` with the `FeeVersion` instead.