Add epoch stake table retrieval support to node-metrics
The node-validator data state does not currently update it's stake
table when a new epoch arrives.  If the stake table changes
between epochs, this will cause the voter participation to be
miscalculated resulting in some odd numbers and behaviors.

This change adds support for stake table changing based on the
epoch number.  Whenever a new block comes through that has
a different epoch than the previous block, the stake table will
be updated, and new node identities will be added.